Repairs `_escape` method on strings.
This method is a hack, and should not be present.
But the C code generaotr relies on it!  This patch
makes it work well enough for the compiler to compile
itself.  This code should be moved into the C code generator,
but I don't want to spend time on that now.